test : 주소 테스트를 위한 테스트 코드

+++ src/main/java/com/goodfood/goodfood/global/repository/RestaurantRepository.java
@@ -3,5 +3,5 @@ package com.goodfood.goodfood.global.repository;
 import com.goodfood.goodfood.global.entity.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
+public interface RestaurantRepository extends JpaRepository<Restaurant, Integer> {
 }
